+++ src/wai/annotations/image_utils/__init__.py
@@ -7,3 +7,5 @@ from ._image import remove_alpha_channel
 from ._masks import mask_to_polygon
 from ._masks import polygon_to_minrect
 from ._masks import polygon_to_lists
+from ._masks import lists_to_polygon
+from ._masks import polygon_to_bbox